refactor: move exported one time functions closer to their user
