+++ apps/main/src/dex/features/pool-list/cells/LegacyRewardsCrvCell.tsx
@@ -4,10 +4,10 @@ import type { CellContext } from '@tanstack/react-table'
 import { isSortedBy } from '@ui-kit/shared/ui/DataTable/data-table.utils'
 import { LegacyPoolColumnId } from '../columns'
 import { hasLegacyCrvRewards } from '../hooks/useLegacyHasPoolRewards'
-import type { LegacyPoolListItem } from '../legacy-pools.types'
+import type { LegacyPoolRow } from '../types'
 import { Placeholder } from './Placeholder'
 
-type Prop = CellContext<LegacyPoolListItem, RewardsApy | undefined> & { placeholder?: boolean }
+type Prop = CellContext<LegacyPoolRow, RewardsApy | undefined> & { placeholder?: boolean }
 
 export const LegacyRewardsCrvCell = ({ getValue, table, row: { original: poolData }, placeholder = true }: Prop) => {
   const rewards = getValue()
