+++ apps/main/src/dex/features/pool-list/cells/LegacyRewardsOtherCell.tsx
@@ -1,12 +1,12 @@
 import { RewardsApy } from '@/dex/types/main.types'
 import type { CellContext } from '@tanstack/react-table'
 import { useLegacyHasPoolRewards } from '../hooks/useLegacyHasPoolRewards'
-import type { LegacyPoolListItem } from '../legacy-pools.types'
+import type { LegacyPoolRow } from '../types'
 import { LegacyRewardsCrvCell } from './LegacyRewardsCrvCell'
 import { LegacyRewardsIncentivesCell } from './LegacyRewardsIncentivesCell'
 import { Placeholder } from './Placeholder'
 
-type Prop = CellContext<LegacyPoolListItem, RewardsApy | undefined>
+type Prop = CellContext<LegacyPoolRow, RewardsApy | undefined>
 
 export const LegacyRewardsOtherCell = (props: Prop) => {
   const rewards = props.getValue()
