+++ apps/main/src/dex/features/pool-list/cells/LegacyRewardsOtherHeader.tsx
@@ -7,11 +7,11 @@ import { t } from '@ui-kit/lib/i18n'
 import { Sortable } from '@ui-kit/shared/ui/DataTable/Sortable'
 import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
 import { LegacyPoolColumnId } from '../columns'
-import type { LegacyPoolListItem } from '../legacy-pools.types'
+import type { LegacyPoolRow } from '../types'
 
 const { Spacing } = SizesAndSpaces
 
-export const LegacyRewardsOtherHeader = ({ table }: HeaderContext<LegacyPoolListItem, RewardsApy | undefined>) => (
+export const LegacyRewardsOtherHeader = ({ table }: HeaderContext<LegacyPoolRow, RewardsApy | undefined>) => (
   <Stack>
     <Box>{t`Rewards tAPR`}</Box>
     <Stack direction="row" sx={{ gap: Spacing.xs, alignItems: 'end' }}>
