+++ apps/main/src/dex/features/pool-list/cells/LegacyUsdCell.tsx
@@ -2,9 +2,9 @@ import Stack from '@mui/material/Stack'
 import type { CellContext } from '@tanstack/react-table'
 import { Tooltip } from '@ui-kit/shared/ui/Tooltip'
 import { formatNumber } from '@ui-kit/utils'
-import type { LegacyPoolListItem } from '../legacy-pools.types'
+import type { LegacyPoolRow } from '../types'
 
-export const LegacyUsdCell = ({ getValue }: CellContext<LegacyPoolListItem, number | null>) => {
+export const LegacyUsdCell = ({ getValue }: CellContext<LegacyPoolRow, number | null>) => {
   const value = getValue()
   return (
     <Tooltip title={value && formatNumber(value, 'usd.amount')}>
