+++ apps/main/src/dex/features/pool-list/cells/PoolTitleCell/LegacyPoolTitleCell.tsx
@@ -10,7 +10,7 @@ import { UserPositionIndicator } from '@ui-kit/shared/ui/DataTable/UserPositionI
 import { TokenIcons } from '@ui-kit/shared/ui/TokenIcons'
 import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
 import { MarketTitle } from '@ui-kit/widgets/MarketTitle'
-import type { LegacyPoolListItem } from '../../legacy-pools.types'
+import type { LegacyPoolRow } from '../../types'
 import { PoolAlertBadge } from './PoolAlertBadge'
 import { PoolAlertIcons } from './PoolAlertIcons'
 import { PoolTokens } from './PoolTokens'
@@ -20,7 +20,7 @@ const { Spacing, Height } = SizesAndSpaces
 export const LegacyPoolTitleCell = ({
   row: { original: poolData },
   column: { getFilterValue },
-}: CellContext<LegacyPoolListItem, string>) => {
+}: CellContext<LegacyPoolRow, string>) => {
   const { tokenAddresses, tokens, tokenAddressesAll, pool, url, network } = poolData
   const tokenList = useMemo(
     () => lodash.zip(tokens, tokenAddresses).map(([symbol, address]) => ({ symbol: symbol!, address: address! })),
