+++ apps/main/src/dex/features/pool-list/components/LegacyPoolExpandedPanelActions.tsx
@@ -3,9 +3,9 @@ import { getPath } from '@/dex/utils/utilsRouter'
 import { t } from '@ui-kit/lib/i18n'
 import { ExpandedPanelActions } from '@ui-kit/shared/ui/DataTable/ExpandedPanelActions'
 import type { ExpandedPanelComponent } from '@ui-kit/shared/ui/DataTable/ExpansionRow'
-import type { LegacyPoolListItem } from '../legacy-pools.types'
+import type { LegacyPoolRow } from '../types'
 
-export const LegacyPoolExpandedPanelActions: ExpandedPanelComponent<LegacyPoolListItem> = ({ row }) => {
+export const LegacyPoolExpandedPanelActions: ExpandedPanelComponent<LegacyPoolRow> = ({ row }) => {
   const {
     pool: { id: poolId },
     network,
