+++ apps/main/src/dex/features/pool-list/components/PoolExpandedPanelActions.tsx
@@ -4,9 +4,9 @@ import { copyToClipboardWithToast } from '@ui-kit/hooks/useCopyToClipboard'
 import { t } from '@ui-kit/lib/i18n'
 import { ExpandedPanelActions } from '@ui-kit/shared/ui/DataTable/ExpandedPanelActions'
 import type { ExpandedPanelComponent } from '@ui-kit/shared/ui/DataTable/ExpansionRow'
-import type { PoolListItem } from '../pools.types'
+import type { PoolRow } from '../types'
 
-export const PoolExpandedPanelActions: ExpandedPanelComponent<PoolListItem> = ({ row }) => {
+export const PoolExpandedPanelActions: ExpandedPanelComponent<PoolRow> = ({ row }) => {
   const pool = row.original
   const path = getPath({ network: pool.network }, `${ROUTE.PAGE_POOLS}/${pool.address}`)
