+++ apps/main/src/dex/features/pool-list/hooks/useLegacyHasPoolRewards.ts
@@ -3,11 +3,11 @@ import type { RewardsApy } from '@/dex/types/main.types'
 import type { Chain } from '@curvefi/prices-api'
 import type { Address } from '@primitives/address.utils'
 import { useCampaignsByAddress } from '@ui-kit/entities/campaigns'
-import type { LegacyPoolListItem } from '../legacy-pools.types'
+import type { LegacyPoolRow } from '../types'
 
 export const hasLegacyCrvRewards = (rewards: RewardsApy | undefined) => sum(rewards?.crv) > 0
 
-export const useLegacyHasPoolRewards = (rewards: RewardsApy | undefined, poolData: LegacyPoolListItem) => {
+export const useLegacyHasPoolRewards = (rewards: RewardsApy | undefined, poolData: LegacyPoolRow) => {
   const { data: campaigns } = useCampaignsByAddress({
     blockchainId: poolData.network as Chain,
     address: poolData?.pool?.address as Address,
